requirements.txt
fastapi==0.115.0
uvicorn==0.31.1

telethon==1.37.0

openai>=1.6.1

tiktoken==0.7.0

python-dotenv==1.0.1
duckduckgo-search==6.3.2
pytz==2024.2
Unidecode==1.3.8
coloredlogs==15.0.1
